backend/requirements.txt
numpy==2.4.3
pydantic==2.12.5
PyYAML==6.0.3
